pyproject.toml
[project]
name = "typhoon-webrtc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.10",
    "aiortc>=1.9.0",
    "av>=12.3.0",
    "librosa>=0.10.2.post1",
    "numpy>=2.0.2",
    "opencv-python-headless>=4.10.0.84",
    "peft>=0.14.0",
    "python-socketio>=5.11.4",
    "sounddevice>=0.5.1",
    "soundfile>=0.12.1",
    "torch>=2.5.1",
    "torchaudio>=2.5.1",
    "transformers>=4.47.0",
]
